// apps/web/src/lib/__generated__/api.ts
export type BatchDtoStatus = (typeof BatchDtoStatus)[keyof typeof BatchDtoStatus];

export const BatchDtoStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type BatchDto = {
  id: string;
  name: string;
  status: BatchDtoStatus;
  totalCount: number;
  succeededCount: number;
  failedCount: number;
  createdAt: string;
  updatedAt: string;
};